clausematch-backend/services/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.9
pdfminer.six==20231228
python-docx==0.8.11
openpyxl==3.1.5
python-pptx==0.6.23
